clang-tidy fixes

+++ shim/src/handle_vcpu_kvm_interrupt.c
@@ -39,13 +39,14 @@
  *
  * <!-- inputs/outputs -->
  *   @param pmut_ioctl_args the arguments provided by userspace
+ *   @param vcpu the struct shim_vcpu_t type to use
  *   @return SHIM_SUCCESS on success, SHIM_FAILURE on failure.
  */
 NODISCARD int64_t
 handle_vcpu_kvm_interrupt(
     struct shim_vcpu_t const *const vcpu, struct kvm_interrupt *const pmut_ioctl_args) NOEXCEPT
 {
-    const uint32_t max_irqs = 256;
+    const uint32_t max_irqs = (uint32_t)256;
 
     if (detect_hypervisor()) {
         bferror("The shim is not running in a VM. Did you forget to start MicroV?");
@@ -59,7 +60,7 @@ handle_vcpu_kvm_interrupt(
         return SHIM_FAILURE;
     }
 
-    if (mv_vs_op_queue_interrupt(g_mut_hndl, vcpu->vsid, pmut_ioctl_args->irq)) {
+    if (mv_vs_op_queue_interrupt(g_mut_hndl, vcpu->vsid, (uint64_t)pmut_ioctl_args->irq)) {
         return SHIM_FAILURE;
     }
